README with link to show --list output (resources/LIST.md)
The dynamic --list command should be the real source of truth (config dependent), but somewhere in the README (docs) we
should show all the licenses we are providing out-of-the-box at least as an example (ideally current default or close to it).

I put --list output under resources/LIST.md (assuming we will add a resources/README.md in the future).

The main README.md just refers to this w/ a one-liner and link.  Probably too subtle, but entire listing takes up a lot of space.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>