Add file to CLI results when using --dir to make the output more usable
The initial implementation of --dir loops and prints for each file like the singular --file.
But since --dir can have a bunch of files, need to print the file name on each "FOUND..." and "No licenses..." line,
so that we can understand the results better (visually and/or by grepping/sorting/counting).

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>