beginOptional text at beginning of line needs to be normalized on newline
Some normalization works on the first non-white-space characters of a line -- e.g. remove leading comment char #.
But a template line starting with <beginOptional># won't get the same treatment and so it will not match.
A reasonable fix for existing known mismatches, is to replace "^<beginOptional>" with "<beginOptional>\n".
Doing this in normalizer allows following normalization steps to recognize that optional block text as a new line.

Fixes: #8

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>