Fix bullet and inline outline numbering replacement and matcher
We had many failed matches due to outline numbering like 1) (vii) and a) which often
appears mid-paragraph in the template but then shows up on a newline in license text.
The newline caused our bullet-handling to remove the numbers from the license while the
template still required them.

This change introduces a new regex to match bullets and numbering separately. They are
similar, but we still have the concept of true bullets always starting a line.

This change also introduces an IsTemplate flag in NormalizationData and a convenience factory
NewNormalizationData(text, flag) with that significant flag.

Now normalization behaves differently for templates because ideally we can remove optional things
from a license, but in these cases it is better to add a matcher to the template instead of just
removing things.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>