Better separation of bullet vs numbered handling. Remove bullets from input.
There was overlap between bullet and numbering.
Both need wildcard matching in templates.
Now we keep the numbers in input text, but remove the bullets.

Numbering is important for whitespace separation and is typically hard-coded in templates.
Bullets are prefixes that are more optional. Sometimes they are in templates and sometimes just added to licenses.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>