Fix SPDX 3.18 matching and update default library (#24)
* Fix bullet and inline outline numbering replacement and matcher

We had many failed matches due to outline numbering like 1) (vii) and a) which often
appears mid-paragraph in the template but then shows up on a newline in license text.
The newline caused our bullet-handling to remove the numbers from the license while the
template still required them.

This change introduces a new regex to match bullets and numbering separately. They are
similar, but we still have the concept of true bullets always starting a line.

This change also introduces an IsTemplate flag in NormalizationData and a convenience factory
NewNormalizationData(text, flag) with that significant flag.

Now normalization behaves differently for templates because ideally we can remove optional things
from a license, but in these cases it is better to add a matcher to the template instead of just
removing things.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>

* Remove NBSP chars during normalization

Added NBSP (U+00A0) to the odd characters list.
This showed up in the IBM-pibs test data (not in the template).
Handling in normalization should be a good fix for here and future cases.
A couple of template prechecks were affected.
Also added a missing build flag and license to test files.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>

* Update full SPDX 3.18 template support

1. SPDX default library updated with all the 3.18 templates.
2. More flexible matching around inconsistent punctuation, bullets, whitespace, tags, etc...

* bullet/numbering handling improved for more test cases
  - regex matching tweaks
  - not removing bullet from license anymore
  - adding wildcard match to templates
* eat more spaces around template <<regex>> matchers
* eat spaces around commas and colons
* added some unicode chars to our whitepace matcher
* using a placeholder symbol diamond for removed tags instead of "" as needed to preserve word boundaries
* Removing "**" mid-line instead of just begin/end so the block comment
  removal won't cause mismatch with line break differences.
* improved removal of HTML tags to avoid corrupting templates
* Varietal word replacement moved to after whitespace reduction
* For now... using replacement_words.json (replaceVarietalWordSpellings() as implementation of whitespace removal around commas, colons, and **
* More importer/validate SPDX template tests that needed fix/test
* Add diff for importer/validate test errors
* Handle deprecated_ SPDX file name mismatch with a retry
* update tests, as needed (i.e. normalized text changes)

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>

* Fix directory test to ignore new /invalid/ testdata files

* This holds files that fail addAll and should not be used for this test.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>

* Better separation of bullet vs numbered handling. Remove bullets from input.

There was overlap between bullet and numbering.
Both need wildcard matching in templates.
Now we keep the numbers in input text, but remove the bullets.

Numbering is important for whitespace separation and is typically hard-coded in templates.
Bullets are prefixes that are more optional. Sometimes they are in templates and sometimes just added to licenses.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>